Client: display something at home page. Move balance sheet to /accounts
