+++ client/components/header.js
@@ -3,6 +3,7 @@ const Header = ({ currentUser }) => {
   const links = [
     !currentUser && { label: 'Sign up', href: '/auth/signup' },
     !currentUser && { label: 'Sign in', href: '/auth/signin' },
+    currentUser && { label: 'Accounts', href: '/accounts' },
     currentUser && { label: 'Transactions', href: '/transactions' },
     currentUser && { label: 'Sign out', href: '/auth/signout' },
   ]
